This is the initial commit for the project! It has a working copy that generates permutations of flight combinations specified in one of the config-files/*.json files. The permutations create flight variations based on the supplied origins, destinations and dates for both depart and returning flights. This means that the initial version supports multi-city destinations of the form: start -> end, another start -> end.
The way to use the tool is to pass a json file as an argument. The file
can follow the same structure as config-files/*.json files. Look at them
for examples.

Currently only Kayak is supported for a website that can be searched.
Later, more websites could be added potentially, but time will show.
Enjoy!

Stanimir Ivanov, Februray 25, 2015